data_generator/requirements.txt
celery
pyyaml
requests